connection: Delete IPv6 default routes
The correct way to delete the default route is to use
__connman_inet_del_default_from_table. connman_inet_clear_ipv6_gateway_address()
will only remove the additional route to reach the gateway on given
interface but not the default route. IPv4 and IPv6 seems to behave
slightly different, as connman_inet_clear_ipv4_gateway_address removes
the default route as well.